Support Zheap in-place update and delete operations
This allows zheap tuples to be updated in-place and marked as deleted.
This also supports visibility checking for snapshots and traversing
the undo chains for non-visible tuples.  As of this commit, the
support for snapshot visibility with respect to command id is not
complete and can give wrong answers.  The future commit will support
the same.