Support reuse of transaction slots
Currently the zheap page has four transaction slots which means that
there can't be more than four active transactions operating on a page.
After four transactions the system will start waiting, this commit
will allow to reuse the transaction slots.  The transaction slot can
reuse if the transaction is committed and all-visible or if it is
aborted then the undo has been applied.  As of now, we don't have
rollback support so we just rely on the first check to reuse the slot.