Implement PRIMARY KEY for zheap relations.
This patch allows PRIMARY KEY to be created on zheap relations.
For now, all it does is, throw an error message, if user tries
to add a NULL or duplicate value on primary-key column (i.e. when
primary key constraint is violated) but it doesn't protect it
from getting inserted into a zheap table as currently ROLLBACK is
not implemented in zheap. Therefore, if users tries to violate a
primary key constraint on zheap table, it will result into a data
corruption.

RM42103, Ashutosh Sharma, reviewed by Kuntal Ghosh and Amit Kapila.
Few adjustments done by Amit Kapila.