WAL log Insert operation in zheap
We need to ensure that we log enough information that during recovery,
we can construct the unpackedundo record and then insert it in same
way as we insert for the DO operation.  Note, that we don't write full
page image for undo logs as those are written serially and we always
apply all the WAL records for undo log unlike for zheap pages where
we don't apply it, if the LSN on a page is greater than the LSN of WAL
record.  This is to avoid problems like if the page header is written
after last checkpoint, but the other part is not then the LSN in page
header will be latest but not all the data.

Patch by me, reviewed by Dilip Kumar.