Bug fix in trigger for zheap
Currently we are always getting the older as well as newer version
of tuple from the zheap. But, in case of inplace update both are
having same tid and will fetch the tuple.  Patch will fix the bug
by getting the older version of the tuple from the undolog.

Patch by me, reviewed by Ashutosh Sharma and Amit Kapila.