WAL log update operation in zheap
The important consideration for logging update operation is that we
need to generate entire old tuple while performing redo operation for
undo log.  So, we are conditionally (when full page writes are off)
writing complete tuple in WAL as well.  In case full page writes are
on, we can rely on the tuple from page. For the new tuple, we are
writing just the diff tuple as we are doing for heap.

Patch by me, reviewed by Kuntal Ghosh.