Recovery of zheap relations when rollbacks are pending
Now, the undo actions are performed at the time of restart
for all those transactions that were in progress at the time
of system crash or the last time when system was up. This
ensures proper recovery of transactions involving zheap relations.

A caveat to note here is that undo worker is initialised with the
connection to default database 'postgres'.

Reviewed by Dilip Kumar and Amit Kapila