Create missing undo log segments during recovery.
During recovery, we might discover that the segment files that existed at the
time of the checkpoint don't exist, because they'll be deleted by later WAL
traffic.  We'll create zero-filled files to avoid errors, and trust that the
contents of the files will never be needed, because later WAL records discard
them.