Remove UndoDiscard and move its state into UndoLogControl.
Previously, a per-undo log object "UndoDiscard" was used to track the progress
of the undo worker machinery.  It was in a shared memory array of fixed size,
which isn't going to work.  Move that state into the UndoLogControl object for
each undo log.

Since this change requires undodiscard.c to have direct access to
UndoLogControl objects and to iterate over them, create new functions
UndoLogNext(), UndoLogGet() with extern linkage to do that.  A better
interface is probably needed -- to review later when we work out the type of
access that multi-process undo worker infrastructure will need.

Register the LWLock tranches.

Thomas Munro, RM43420, reviewed by Dilip Kumar and Amit Kapila