Two pass vacuum
We need vacuum in zheap for non-delete marked indexes, however, we can
use undo to reduce three-pass to two-pass vacuum.  When a row is
deleted, the vacuum will directly mark the line pointer as unused,
writing an undo record as it does, and then mark the corresponding
index entries as dead.  If vacuum fails midway through the undo can
ensure that changes to the heap page are rolled back.  If the vacuum
goes on to commit, we don't need to revisit the heap page after index
cleanup.

We must be careful about  TID reuse: we will only allow a TID to be
reused when the transaction that has marked it as unused has
committed. At that point, we can be assured that all the index
entries corresponding to dead tuples will be marked as dead.

Currently, due to lack of visibility map for zheap, we scan all the
pages during vacuum.  A future patch which will introduce visibility
map in zheap will remove that limitation.

Patch by me, Mithun has fixed few bugs and added implementation for
Rollback action, also he has done basic verification of the patch