Additional flag in execute_undo_actions for relation lock
Callers of execute_undo_actions now provide a flag based on if they
a lock on the relation. If the caller have relation lock already
then no need to lock again in execute_undo_actions.

This is required particularly in cases when rollbacking the prepared
transactions or rollback to savepoints. In such cases the transaction
have already held the relation locks and we need not take another.

Reviewed by Dilip Kumar