Avoid using tuple freed by a visibility function.
In zheap_get_latest_tid(), the tuple that is passed to
ZHeapTupleSatisfiesVisibility() was being used subsequently, ignoring
the fact that the visibility function frees the passed-in tuple.

Rather than the passed-in tuple, use the tuple returned by the
visibility function in the subsequent code. While at it, make sure
that the returned tuple is also freed if it is different than the
passed-in tuple.

Discovered by Neha Sharma while testing TidScan implementation for
zheap.

Amit Khandekar, reviewed by Kuntal Ghosh.