Implement stats for in-place updates
Reuse the existing hot-update stat variable for in-place updates
instead of introducing a new variable as that will increase the size
of stats structure.  It appears ugly to overload the unrelated
variable, but it is better to discuss in community before introducing
a new variable.

Beena Emerson, reviewed by Mithun C Y and me